# Conflicts:
#	Rock/Rock.csproj
#	RockWeb/Blocks/Core/BlockProperties.ascx.cs

+++ RockWeb/Blocks/WorkFlow/WorkflowTypeDetail.ascx.cs
@@ -19,6 +19,7 @@ using System.Collections.Generic;
 using System.ComponentModel;
 using System.Data.Entity;
 using System.Linq;
+using System.Reflection;
 using System.Web.UI;
 
 using Newtonsoft.Json;
@@ -33,6 +34,7 @@ using Rock.Utility;
 using Rock.Web.Cache;
 using Rock.Web.UI;
 using Rock.Web.UI.Controls;
+using Rock.Workflow;
 using Attribute = Rock.Model.Attribute;
 
 namespace RockWeb.Blocks.WorkFlow
@@ -459,8 +461,27 @@ This {{ Workflow.WorkflowType.WorkTerm }} does not currently require your attent
                                     newFormAttribute.Attribute = new Rock.Model.Attribute
                                     {
                                         Guid = guidXref[formAttribute.Attribute.Guid],
-                                        Name = formAttribute.Attribute.Name
+                                        Name = formAttribute.Attribute.Name,
+                                        FieldTypeId = formAttribute.Attribute.FieldTypeId,
+                                        AttributeQualifiers = formAttribute.Attribute.AttributeQualifiers
                                     };
+
+                                    if ( newFormAttribute.FieldVisibilityRules != null )
+                                    {
+                                        var visibilityRules = newFormAttribute.FieldVisibilityRules.Clone();
+
+                                        foreach ( var rule in visibilityRules.RuleList )
+                                        {
+                                            if ( rule.ComparedToFormFieldGuid != null && guidXref.ContainsKey( rule.ComparedToFormFieldGuid.Value ) )
+                                            {
+                                                rule.ComparedToFormFieldGuid = guidXref[rule.ComparedToFormFieldGuid.Value];
+                                            }
+                                        }
+
+                                        newFormAttribute.FieldVisibilityRules = visibilityRules;
+                                    }
+
+
                                     newWorkflowForm.FormAttributes.Add( newFormAttribute );
                                 }
                             }
@@ -599,6 +620,7 @@ This {{ Workflow.WorkflowType.WorkTerm }} does not currently require your attent
                 workflowType.ProcessingIntervalSeconds = null;
             }
 
+            workflowType.MaxWorkflowAgeDays = tbMaximumWorkflowAge.Text.AsIntegerOrNull();
             workflowType.LogRetentionPeriod = tbLogRetention.Text.AsIntegerOrNull();
             workflowType.CompletedWorkflowRetentionPeriod = tbCompletedRetention.Text.AsIntegerOrNull();
             workflowType.IsPersisted = cbIsPersisted.Checked;
@@ -718,6 +740,8 @@ This {{ Workflow.WorkflowType.WorkTerm }} does not currently require your attent
             }
             rockContext.SaveChanges();
 
+            var formBuilderEntityTypeId = EntityTypeCache.GetId<Rock.Workflow.Action.FormBuilder>();
+
             // add or update WorkflowActivityTypes(and Actions) that are assigned in the UI
             int workflowActivityTypeOrder = 0;
             foreach ( var editorWorkflowActivityType in ActivityTypesState )
@@ -745,12 +769,6 @@ This {{ Workflow.WorkflowType.WorkTerm }} does not currently require your attent
                     SaveAttributes( new WorkflowActivity().TypeId, "ActivityTypeId", workflowActivityType.Id.ToString(), ActivityAttributesState[workflowActivityType.Guid], rockContext );
                 }
 
-                // Because the SaveAttributes above may have flushed the cached entity attribute cache, and it would get loaded again with
-                // a different context, manually reload the cache now with our context to prevent a database lock conflict (when database is 
-                // configured without snapshot isolation turned on)
-                EntityAttributesCache.Remove();
-                EntityAttributesCache.Get( rockContext );
-
                 int workflowActionTypeOrder = 0;
                 foreach ( var editorWorkflowActionType in editorWorkflowActivityType.ActionTypes )
                 {
@@ -773,14 +791,16 @@ This {{ Workflow.WorkflowType.WorkTerm }} does not currently require your attent
                     workflowActionType.AttributeValues = editorWorkflowActionType.AttributeValues;
                     workflowActionType.Order = workflowActionTypeOrder++;
 
-                    if ( workflowActionType.WorkflowForm != null && editorWorkflowActionType.WorkflowForm == null )
+                    var isFormBuilderAction = workflowActionType.EntityTypeId == formBuilderEntityTypeId;
+
+                    if ( !isFormBuilderAction && workflowActionType.WorkflowForm != null && editorWorkflowActionType.WorkflowForm == null )
                     {
                         // Form removed
                         workflowFormService.Delete( workflowActionType.WorkflowForm );
                         workflowActionType.WorkflowForm = null;
                     }
 
-                    if ( editorWorkflowActionType.WorkflowForm != null )
+                    if ( !isFormBuilderAction && editorWorkflowActionType.WorkflowForm != null )
                     {
                         if ( workflowActionType.WorkflowForm == null )
                         {
@@ -823,6 +843,7 @@ This {{ Workflow.WorkflowType.WorkTerm }} does not currently require your attent
                             formAttribute.HideLabel = editorAttribute.HideLabel;
                             formAttribute.PreHtml = editorAttribute.PreHtml;
                             formAttribute.PostHtml = editorAttribute.PostHtml;
+                            formAttribute.FieldVisibilityRules = editorAttribute.FieldVisibilityRules;
                         }
                     }
                 }
@@ -901,7 +922,7 @@ This {{ Workflow.WorkflowType.WorkTerm }} does not currently require your attent
         {
             ParseControls();
 
-            Guid attributeGuid = (Guid)e.RowKeyValue;
+            Guid attributeGuid = ( Guid ) e.RowKeyValue;
             ShowAttributeEdit( attributeGuid );
         }
 
@@ -930,7 +951,7 @@ This {{ Workflow.WorkflowType.WorkTerm }} does not currently require your attent
         {
             ParseControls();
 
-            Guid attributeGuid = (Guid)e.RowKeyValue;
+            Guid attributeGuid = ( Guid ) e.RowKeyValue;
             AttributesState.RemoveEntity( attributeGuid );
 
             BindAttributesGrid();
@@ -1286,7 +1307,7 @@ This {{ Workflow.WorkflowType.WorkTerm }} does not currently require your attent
 
             if ( workflowType.IsSystem )
             {
-                readOnly = true; 
+                readOnly = true;
                 nbEditModeMessage.Heading = "Information";
                 nbEditModeMessage.Text = EditModeMessage.ReadOnlySystem( WorkflowType.FriendlyTypeName );
             }
@@ -1410,6 +1431,11 @@ This {{ Workflow.WorkflowType.WorkTerm }} does not currently require your attent
             }
 
             SetEditMode( true );
+            nbWarningMessage.Visible = workflowType.IsFormBuilder;
+            if ( workflowType.IsFormBuilder )
+            {
+                nbWarningMessage.Text = "This workflow was created from the Form Builder feature. Care should be taken in editing this as a workflow to ensure that you do not break the native capabilities.";
+            }
 
             cbIsActive.Checked = workflowType.IsActive == true;
             tbName.Text = workflowType.Name;
@@ -1428,10 +1454,11 @@ This {{ Workflow.WorkflowType.WorkTerm }} does not currently require your attent
                 tbProcessingInterval.Text = string.Empty;
             }
 
+            tbMaximumWorkflowAge.Text = workflowType.MaxWorkflowAgeDays.ToStringSafe();
             tbLogRetention.Text = workflowType.LogRetentionPeriod.ToStringSafe();
             tbCompletedRetention.Text = workflowType.CompletedWorkflowRetentionPeriod.ToStringSafe();
             cbIsPersisted.Checked = workflowType.IsPersisted;
-            ddlLoggingLevel.SetValue( (int)workflowType.LoggingLevel );
+            ddlLoggingLevel.SetValue( ( int ) workflowType.LoggingLevel );
             tbIconCssClass.Text = workflowType.IconCssClass;
             ceSummaryViewText.Text = workflowType.SummaryViewText;
             ceNoActionMessage.Text = workflowType.NoActionMessage;
@@ -1670,6 +1697,7 @@ This {{ Workflow.WorkflowType.WorkTerm }} does not currently require your attent
                 bool showInvalid = false )
         {
             var control = new WorkflowActionTypeEditor();
+            control.IsEditable = DetermineIfActionIsEditable( actionType );
             parentControl.Controls.Add( control );
             control.ID = actionType.Guid.ToString( "N" );
             control.ValidationGroup = btnSave.ValidationGroup;
@@ -1695,10 +1723,11 @@ This {{ Workflow.WorkflowType.WorkTerm }} does not currently require your attent
                 // Add any new attributes
                 foreach ( var attribute in attributes )
                 {
-                    if ( !formAttributes.Select( a => a.Attribute.Guid ).Contains( attribute.Key ) )
+                    var formAttribute = formAttributes.Where( fa => fa.Attribute.Guid == attribute.Key ).FirstOrDefault();
+                    if ( formAttribute == null )
                     {
-                        var formAttribute = new WorkflowActionFormAttribute();
-                        formAttribute.Attribute = new Rock.Model.Attribute { Guid = attribute.Key, Name = attribute.Value.Name };
+                        formAttribute = new WorkflowActionFormAttribute();
+                        formAttribute.Attribute = new Attribute { Guid = attribute.Key, Name = attribute.Value.Name, FieldTypeId = attribute.Value.FieldTypeId };
                         formAttribute.Guid = Guid.NewGuid();
                         formAttribute.Order = formAttributes.Any() ? formAttributes.Max( a => a.Order ) + 1 : 0;
                         formAttribute.IsVisible = false;
@@ -1709,6 +1738,9 @@ This {{ Workflow.WorkflowType.WorkTerm }} does not currently require your attent
                         formAttribute.PostHtml = string.Empty;
                         formAttributes.Add( formAttribute );
                     }
+
+                    // make sure formAttribute AttributeQualifiers matches the current qualifiers if they have been added/edited
+                    formAttribute.Attribute.AttributeQualifiers = attribute.Value.AttributeQualifiers.ToList();
                 }
             }
             control.SetWorkflowActionType( actionType, attributes );
@@ -1737,6 +1769,30 @@ This {{ Workflow.WorkflowType.WorkTerm }} does not currently require your attent
             }
 
             return control;
+        }
+
+        /// <summary>
+        /// Determines if action type is editable.
+        /// </summary>
+        /// <param name="actionType">Type of the action.</param>
+        private bool DetermineIfActionIsEditable( WorkflowActionType actionType )
+        {
+            ActionComponent actionComponent = actionType?.WorkflowAction;
+
+            if ( actionComponent == null )
+            {
+                // probably adding a new workflow action and type hasn't been selected yet
+                return true;
+            }
+
+            var attr = actionComponent.GetType().GetCustomAttribute<ActionCategoryAttribute>( false );
+
+            if ( attr != null && attr.CategoryName == "HideFromUser" )
+            {
+                return false;
+            }
+
+            return true;
         }
 
         /// <summary>
